some rearranging to separate the two MCP servers (weather and hackernews)
