+++ workflows/hackernews_activities.py
@@ -1,27 +1,10 @@
-# activities.py
+# hackernews_activities.py
 
 from typing import Any
 from temporalio import activity
 import httpx
 from shared.models import HackerNewsParams
 from typing import Optional
-import asyncio
-
-USER_AGENT = "weather-app/1.0"
-
-# External calls happen via activities now
-@activity.defn
-async def make_nws_request(url: str) -> dict[str, Any] | None:
-    """Make a request to the NWS API with proper error handling."""
-    headers = {
-        "User-Agent": USER_AGENT,
-        "Accept": "application/geo+json"
-    }
-    async with httpx.AsyncClient() as client:
-
-        response = await client.get(url, headers=headers, timeout=5.0)
-        response.raise_for_status()
-        return response.json()
 
 @activity.defn
 async def make_hackernews_request(params: HackerNewsParams) -> dict[str, Any] | None:
